Move UbloxFirmware8 class to its own files.
